% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_crosstabs.R
\name{cross_tabulate}
\alias{cross_tabulate}
\title{cross_tabulate}
\usage{
cross_tabulate(data, rows, columns, by_row = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{rows}{the variable for the rows, as a string}

\item{columns}{the variable for the columns, as a string}
}
\value{
dataframe
}
\description{
Return a 2 x 2 cross tabs from a dataframe with rows and columns
given as parameters
}
